Camera as SDL_FRect; L03: TODO 1/2/3 solution (frame cap, dt in seconds, camera*dt)
Render::camera becomes SDL_FRect instead of SDL_Rect -- SDL3's renderer
is float-native, so the old int camera forced a float -> int -> float
round-trip on every move and truncated sub-pixel motion. DrawTexture,
DrawRectangle, DrawLine and DrawCircle now do their camera-relative
maths entirely in float and round only the final destination once,
last (std::round on the combined camera + world position, not the
camera alone) -- rounding only the camera isn't enough, since entities
at non-integer positions would still land between pixels.

Window::scale and Render's local scale become float to match.

Textures::LoadSurface sets SDL_SCALEMODE_NEAREST right after creating
each texture, since SDL3 defaults to linear filtering, which blurs
pixel art and bleeds colour between adjacent tiles in a tileset.

L03 TODOs filled in:
- TODO 1/2: FinishUpdate() caps each frame to 1000/targetFrameRate ms,
  timing the wait with PerfTimer and delaying with SDL_DelayPrecise.
- TODO 3: Scene::Update() moves the camera by camSpeed * dt (200 px/s)
  straight into the now-float render->camera.x/y.
dt is stored in seconds (matching the L02 decision) and the window
title displays dt * 1000.0 with an explicit "ms" label. targetFrameRate
is a float FPS count, commented accordingly.

Carried forward from L01_Simple_Engine / L02_Framerate_Solution /
L03_DeltaTime. Squashed per the 2026-09 one-time history-rewrite
waiver (course not yet started, nobody has cloned): was cd4d0cf +
261808d + 124556a + b7e0b58 + 70aa1ee + 0124c57.

+++ src/Window.cpp
@@ -36,7 +36,7 @@ bool Window::Awake()
 		// TODO Get the values from the config file
 		width = 1280;
 		height = 720;
-		scale = 1;
+		scale = 1.0f;
 
 		if (fullscreen == true)        flags |= SDL_WINDOW_FULLSCREEN;
 		if (borderless == true)        flags |= SDL_WINDOW_BORDERLESS;
@@ -94,7 +94,7 @@ void Window::GetWindowSize(int& width, int& height) const
 	height = this->height;
 }
 
-int Window::GetScale() const
+float Window::GetScale() const
 {
 	return scale;
 }
